pyproject.toml
[tool.poetry]
name = "forcuanteller"
version = "0.1.0"
description = "Technical Analysis Alert send to your devices!"
authors = ["Aditya Sidharta <aditya.sdrt@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
requests = "^2.25.1"
yfinance = "^0.1.55"
ta = "^0.7.0"
ipykernel = "^5.4.3"
seaborn = "^0.11.1"
matplotlib = "^3.3.4"
pandas = "^1.2.2"
numpy = "^1.20.1"
loguru = "^0.5.3"
oyaml = "^1.0"
plotly = "^4.14.3"
mplfinance = "^0.12.7-alpha.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
